[loader] Move PInvoke and DllMap code into separate source file
