+++ mono/metadata/loader-internals.h
@@ -16,6 +16,15 @@
 typedef struct _MonoLoadedImages MonoLoadedImages;
 typedef struct _MonoAssemblyLoadContext MonoAssemblyLoadContext;
 
+typedef struct _MonoDllMap MonoDllMap;
+struct _MonoDllMap {
+	char *dll;
+	char *target;
+	char *func;
+	char *target_func;
+	MonoDllMap *next;
+};
+
 #ifdef ENABLE_NETCORE
 /* FIXME: this probably belongs somewhere else */
 struct _MonoAssemblyLoadContext {
@@ -30,8 +39,23 @@ struct _MonoAssemblyLoadContext {
 };
 #endif /* ENABLE_NETCORE */
 
+void
+mono_global_loader_data_lock (void);
+
+void
+mono_global_loader_data_unlock (void);
+
 gpointer
 mono_lookup_pinvoke_call_internal (MonoMethod *method, MonoError *error);
+
+void
+mono_dllmap_insert_internal (MonoImage *assembly, const char *dll, const char *func, const char *tdll, const char *tfunc);
+
+void
+mono_global_dllmap_cleanup (void);
+
+void
+mono_cached_module_cleanup (void);
 
 #ifdef ENABLE_NETCORE
 void
