+++ mono/metadata/loader.h
@@ -68,7 +68,7 @@ mono_lookup_internal_call (MonoMethod *method);
 MONO_API const char*
 mono_lookup_icall_symbol (MonoMethod *m);
 
-MONO_API void
+MONO_API MONO_RT_EXTERNAL_ONLY void
 mono_dllmap_insert (MonoImage *assembly, const char *dll, const char *func, const char *tdll, const char *tfunc);
 
 MONO_API MONO_RT_EXTERNAL_ONLY void*
