+++ msvc/libmonoruntime-common.targets
@@ -106,6 +106,7 @@
     </ClCompile>
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\mono-route.h" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\monitor.h" />
+    <ClCompile Include="$(MonoSourceLocation)\mono\metadata\native-library.c" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\normalization-tables.h" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\number-formatter.h" />
     <ClInclude Include="$(MonoSourceLocation)\mono\metadata\number-ms.h" />
